Fix errors related to helpers
